Trip Planner agent/requirements.txt
phidata
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
exa_py

